修改了卡片布局的阴影值，修改了菜谱大全的背景色，修改了访问菜谱详情接口的参数类型。

+++ app/src/main/java/com/hnqcgc/redfirecookbook/logic/Repository.java
@@ -46,7 +46,7 @@ public class Repository {
         return recipeLiveData;
     }
 
-    public MutableLiveData<RecipeDetails> searchRecipe(long recipeId) {
+    public MutableLiveData<RecipeDetails> searchRecipe(int recipeId) {
         MutableLiveData<RecipeDetails> recipeDetailsLiveData = new MutableLiveData<>();
         Call<RecipeDetails> call = RedFireCookBookNetwork.getInstance().searchRecipe(recipeId);
         call.enqueue(new Callback<RecipeDetails>() {
